Move code to pairbot module
