+++ pairbot/utils.py
